address pr concern
