+++ frontend/src/Types/application.ts
@@ -1,6 +1,7 @@
+import { BLOOD_COMPONENTS } from './bloodProducts';
+
 // Time formatting ------------------------------------------------
 // Time period types
-
 export type Quarter = `${number}-Q${1 | 2 | 3 | 4}`;
 export type Month = `${number}-${string}`; // e.g. "2023-Jan"
 export type Year = `${number}`; // e.g. "2023"
@@ -12,107 +13,6 @@ export const AGGREGATION_OPTIONS = {
   avg: { label: 'Average' },
 } as const;
 
-// Blood components -----------------------------------------------
-const BLOOD_COMPONENT_DECIMALS = { sum: 0, avg: 2 };
-
-export const BLOOD_COMPONENTS = [
-  {
-    value: 'rbc_units',
-    label: {
-      short: 'RBCs',
-      base: 'RBCs Transfused',
-      sum: 'Total RBCs Transfused',
-      avg: 'Average RBCs Transfused Per Visit',
-    },
-    units: {
-      sum: 'RBC Units',
-      avg: 'RBC Units Per Visit',
-      sumShort: 'Units',
-      avgShort: 'Units',
-    },
-    decimals: BLOOD_COMPONENT_DECIMALS,
-  },
-  {
-    value: 'ffp_units',
-    label: {
-      short: 'FFP',
-      base: 'FFP Transfused',
-      sum: 'Total FFP Transfused',
-      avg: 'Average FFP Transfused Per Visit',
-    },
-    units: {
-      sum: 'Plasma Units',
-      avg: 'Plasma Units Per Visit',
-      sumShort: 'Units',
-      avgShort: 'Units',
-    },
-    decimals: BLOOD_COMPONENT_DECIMALS,
-  },
-  {
-    value: 'plt_units',
-    label: {
-      short: 'Platelets',
-      base: 'Platelets Transfused',
-      sum: 'Total Platelets Transfused',
-      avg: 'Average Platelets Transfused Per Visit',
-    },
-    units: {
-      sum: 'Platelet Units',
-      avg: 'Platelet Units Per Visit',
-      sumShort: 'Units',
-      avgShort: 'Units',
-    },
-    decimals: BLOOD_COMPONENT_DECIMALS,
-  },
-  {
-    value: 'cryo_units',
-    label: {
-      short: 'Cryo',
-      base: 'Cryo Transfused',
-      sum: 'Total Cryo Transfused',
-      avg: 'Average Cryo Transfused Per Visit',
-    },
-    units: {
-      sum: 'Cryo Units',
-      avg: 'Cryo Units Per Visit',
-      sumShort: 'Units',
-      avgShort: 'Units',
-    },
-    decimals: BLOOD_COMPONENT_DECIMALS,
-  },
-  {
-    value: 'whole_units',
-    label: {
-      short: 'Whole Blood',
-      base: 'Whole Blood Transfused',
-      sum: 'Total Whole Blood Transfused',
-      avg: 'Average Whole Blood Transfused Per Visit',
-    },
-    units: {
-      sum: 'Whole Blood Units',
-      avg: 'Whole Blood Units Per Visit',
-      sumShort: 'Units',
-      avgShort: 'Units',
-    },
-    decimals: BLOOD_COMPONENT_DECIMALS,
-  },
-  {
-    value: 'cell_saver_ml',
-    label: {
-      short: 'Cell Salvage',
-      base: 'Cell Salvage Volume (ml) Used',
-      sum: 'Total Cell Salvage Volume (ml) Used',
-      avg: 'Average Cell Salvage Volume (ml) Used Per Visit',
-    },
-    units: {
-      sum: 'mL',
-      avg: 'mL Per Visit',
-      sumShort: 'mL',
-      avgShort: 'mL',
-    },
-    decimals: BLOOD_COMPONENT_DECIMALS,
-  },
-] as const;
 // Values of blood components
 export type BloodComponent = typeof BLOOD_COMPONENTS[number]['value'];
 // Readonly array of blood component options
